move all path magic into AspNet5Paths
